fixup! Split out reflection inference into separate module
