+++ ktor-shared/ktor-openapi-schema/ktor-openapi-schema-reflect/jvm/test/io/ktor/openapi/reflect/KotlinxJsonSchemaInferenceTest.kt
@@ -2,6 +2,8 @@
  * Copyright 2014-2026 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.openapi
+package io.ktor.openapi.reflect
+
+import io.ktor.openapi.KotlinxJsonSchemaInference
 
 class KotlinxJsonSchemaInferenceTest : AbstractSchemaInferenceTest(KotlinxJsonSchemaInference)
